+++ packages/gov-ui-kit/src/core/components/forms/checkboxGroup/checkboxGroup.tsx
@@ -1,10 +1,28 @@
 import classNames from 'classnames';
-import type { ComponentProps } from 'react';
+import type { ReactNode } from 'react';
+import { useRandomId } from '../../../hooks';
+import { InputContainer, type IInputContainerBaseProps } from '../inputContainer';
 
-export interface ICheckboxGroupProps extends ComponentProps<'div'> {}
+export interface ICheckboxGroupProps
+    extends Pick<IInputContainerBaseProps, 'alert' | 'label' | 'helpText' | 'isOptional'> {
+    /**
+     * Additional classes for the component.
+     */
+    className?: string;
+    /**
+     * Children of the component.
+     */
+    children?: ReactNode;
+}
 
 export const CheckboxGroup: React.FC<ICheckboxGroupProps> = (props) => {
-    const { className, ...otherProps } = props;
+    const { className, children, ...otherProps } = props;
 
-    return <div className={classNames('flex min-w-0 flex-col gap-2 md:gap-3', className)} {...otherProps} />;
+    const id = useRandomId();
+
+    return (
+        <InputContainer id={id} useCustomWrapper={true} {...otherProps}>
+            <div className={classNames('flex min-w-0 flex-col gap-2 md:gap-3', className)}>{children}</div>
+        </InputContainer>
+    );
 };