+++ packages/gov-ui-kit/src/core/components/forms/radioGroup/index.ts
@@ -1,3 +1,1 @@
-export * from './radio';
-export * from './radioCard';
-export * from './radioGroup';
+export { RadioGroup, type IRadioGroupProps } from './radioGroup';