+++ packages/gov-ui-kit/src/core/hooks/useRandomId/useRandomId.ts
@@ -1,0 +1,7 @@
+import { useId } from 'react';
+
+export const useRandomId = (id?: string) => {
+    const randomId = useId();
+
+    return id ?? randomId;
+};